// lib/earth3d/proto/rocktree.ts
export interface INodeKey {
    path?: string;
    epoch?: number;
}

export interface ICopyrightRequest {
    epoch?: number;
}

export interface ITextureDataRequest {
    node_key?: INodeKey;
    texture_format?: Texture_Format_Value;
    view_direction?: Texture_ViewDirection_Value;
}

export interface IBulkMetadata {
    node_metadata?: INodeMetadata[];
    head_node_key?: INodeKey;
    head_node_center?: number[];
    meters_per_texel?: number[];
    default_imagery_epoch?: number;
    default_available_texture_formats?: number;
    default_available_view_dependent_textures?: number;
    default_available_view_dependent_texture_formats?: number;
}

export interface INodeMetadata {
    path_and_flags?: number;
    epoch?: number;
    bulk_metadata_epoch?: number;
    oriented_bounding_box?: Uint8Array;
    meters_per_texel?: number;
    processing_oriented_bounding_box?: number[];
    imagery_epoch?: number;
    available_texture_formats?: number;
    available_view_dependent_textures?: number;
    available_view_dependent_texture_formats?: number;
}

export type NodeMetadata_Flags_Key = "RICH3D_LEAF" | "RICH3D_NODATA" | "LEAF" | "NODATA" | "USE_IMAGERY_EPOCH";
export type NodeMetadata_Flags_Value = 1 | 2 | 4 | 8 | 16;
// ...
export const NodeMetadata = {

    read(pbf: Pbf, end?: number): INodeMetadata {
        return pbf.readFields(NodeMetadata._readField, {path_and_flags: 0, epoch: 0, bulk_metadata_epoch: 0, oriented_bounding_box: null, meters_per_texel: 0, processing_oriented_bounding_box: [], imagery_epoch: 0, available_texture_formats: 0, available_view_dependent_textures: 0, available_view_dependent_texture_formats: 0}, end);
    },
    _readField(tag: number, obj: any, pbf: Pbf): void {
        if (tag === 1) obj.path_and_flags = pbf.readVarint();
        else if (tag === 2) obj.epoch = pbf.readVarint();
        else if (tag === 5) obj.bulk_metadata_epoch = pbf.readVarint();
        else if (tag === 3) obj.oriented_bounding_box = pbf.readBytes();
        else if (tag === 4) obj.meters_per_texel = pbf.readFloat();
        else if (tag === 6) pbf.readPackedDouble(obj.processing_oriented_bounding_box);
        else if (tag === 7) obj.imagery_epoch = pbf.readVarint();
        else if (tag === 8) obj.available_texture_formats = pbf.readVarint();
        else if (tag === 9) obj.available_view_dependent_textures = pbf.readVarint();
        else if (tag === 10) obj.available_view_dependent_texture_formats = pbf.readVarint();
    },
    write(obj: INodeMetadata, pbf: Pbf): void {
        if (obj.path_and_flags) pbf.writeVarintField(1, obj.path_and_flags);
        if (obj.epoch) pbf.writeVarintField(2, obj.epoch);
        if (obj.bulk_metadata_epoch) pbf.writeVarintField(5, obj.bulk_metadata_epoch);
        if (obj.oriented_bounding_box) pbf.writeBytesField(3, obj.oriented_bounding_box);
        if (obj.meters_per_texel) pbf.writeFloatField(4, obj.meters_per_texel);
        if (obj.processing_oriented_bounding_box) pbf.writePackedDouble(6, obj.processing_oriented_bounding_box);
        if (obj.imagery_epoch) pbf.writeVarintField(7, obj.imagery_epoch);
        if (obj.available_texture_formats) pbf.writeVarintField(8, obj.available_texture_formats);
        if (obj.available_view_dependent_textures) pbf.writeVarintField(9, obj.available_view_dependent_textures);
        if (obj.available_view_dependent_texture_formats) pbf.writeVarintField(10, obj.available_view_dependent_texture_formats);
    },

    Flags: {
    "RICH3D_LEAF": {
        "value": 1,
        "options": {}
    },
    "RICH3D_NODATA": {
        "value": 2,
        "options": {}
    },
    "LEAF": {
        "value": 4,
        "options": {}
    },
    "NODATA": {
        "value": 8,
        "options": {}
    },
    "USE_IMAGERY_EPOCH": {
        "value": 16,
        "options": {}
    }
} as { [K in NodeMetadata_Flags_Key]: { value: NodeMetadata_Flags_Value, options: any } }


};